feat: add tsconfig and tsconfig noEmit variants
This adds two new options.

TSConfig and TSConfig noEmit. The first is when writing TypeScript without a bundler and the second when you need a tsconfig but you are also using a bundler.
It is likely that the second scenario will be less needed by users of Project Calabera, but it does not add much of a maintenance burden so I added it.

fix #5